+++ src/ui/pages/diagnostics-detail.tsx
@@ -17,148 +17,8 @@ import {
   IconSource,
   IconInfo,
 } from "../shared/icons";
-import {
-  CategoryScale,
-  Chart as ChartJS,
-  Colors,
-  Legend,
-  LineElement,
-  LinearScale,
-  PointElement,
-  TimeScale,
-  type TimeUnit,
-  Title,
-  Tooltip,
-} from "chart.js";
-import "chartjs-adapter-luxon";
-import { Line } from "react-chartjs-2";
 import { StreamingText } from "../shared/llm";
-
-// Chart.js plugin to draw a vertical line on hover
-const verticalLinePlugin = {
-  id: 'verticalLine',
-  beforeDraw: (chart: ChartJS) => {
-    if (chart.tooltip?.getActiveElements()?.length) {
-      const activePoint = chart.tooltip.getActiveElements()[0];
-      const ctx = chart.ctx;
-      const x = activePoint.element.x;
-      const topY = chart.scales.y.top;
-      const bottomY = chart.scales.y.bottom;
-
-      ctx.save();
-      ctx.beginPath();
-      ctx.moveTo(x, topY);
-      ctx.lineTo(x, bottomY);
-      ctx.lineWidth = 1;
-      ctx.strokeStyle = '#94a3b8';
-      ctx.setLineDash([5, 5]);
-      ctx.stroke();
-      ctx.restore();
-    }
-  }
-};
-
-// ChartJS plugin to draw annotations
-declare module 'chart.js' {
-  interface Chart {
-    annotationAreas?: Array<{
-      x1: number;
-      x2: number;
-      y1: number;
-      y2: number;
-      description: string;
-    }>;
-  }
-
-  interface PluginOptionsByType<TType> {
-    annotations?: Annotation[];
-  }
-}
-
-const annotationsPlugin = {
-  id: 'annotations',
-  afterDraw: (chart: ChartJS, args: any, options: any) => {
-    const ctx = chart.ctx;
-    const annotations = chart.options?.plugins?.annotations || [];
-
-    annotations.forEach((annotation: any) => {
-      // Convert timestamps to numbers for the time scale
-      const xScale = chart.scales.x;
-      const yScale = chart.scales.y;
-
-      // Parse the timestamps into Date objects and get their timestamps
-      const x1 = new Date(annotation.x_min).getTime();
-      const x2 = new Date(annotation.x_max).getTime();
-
-      // Convert to pixel coordinates
-      const pixelX1 = xScale.getPixelForValue(x1);
-      const pixelX2 = xScale.getPixelForValue(x2);
-      const pixelY1 = yScale.getPixelForValue(annotation.y_max);
-      const pixelY2 = yScale.getPixelForValue(annotation.y_min);
-
-      // Draw annotation rectangle
-      ctx.save();
-      ctx.fillStyle = 'rgba(255, 0, 0, 0.5)'; // Solid red with 50% opacity
-      ctx.fillRect(pixelX1, pixelY1, pixelX2 - pixelX1, pixelY2 - pixelY1);
-
-      // Rectangle border
-      ctx.strokeStyle = 'rgb(200, 0, 0)'; // Solid darker red
-      ctx.strokeRect(pixelX1, pixelY1, pixelX2 - pixelX1, pixelY2 - pixelY1);
-
-      // Annotation label
-      ctx.save();
-      const padding = 4;
-      ctx.font = '10px monospace'; // Reduced font size
-      const textMetrics = ctx.measureText(annotation.label);
-      const textHeight = 12; // Reduced height to match smaller font
-      const radius = 4; // Border radius
-
-      // Calculate label box dimensions
-      const boxX = pixelX1 + padding;
-      const boxY = pixelY1 - textHeight - padding * 2;
-      const boxWidth = textMetrics.width + padding * 2;
-      const boxHeight = textHeight + padding * 2;
-
-      // Label box shadow
-      ctx.shadowColor = 'rgba(0, 0, 0, 0.3)';
-      ctx.shadowBlur = 4;
-      ctx.shadowOffsetX = 2;
-      ctx.shadowOffsetY = 2;
-
-      // Label box background
-      ctx.fillStyle = 'rgba(200, 0, 0, 0.75)';
-      ctx.beginPath();
-      ctx.roundRect(boxX, boxY, boxWidth, boxHeight, radius);
-      ctx.fill();
-
-      // Border
-      ctx.shadowColor = 'transparent';
-      ctx.strokeStyle = 'rgb(200, 0, 0)';
-      ctx.lineWidth = 1;
-      ctx.stroke();
-
-      // Label text
-      ctx.fillStyle = 'white';
-      ctx.textBaseline = 'bottom';
-      ctx.fillText(annotation.label, pixelX1 + padding * 2, pixelY1 - padding);
-      ctx.restore();
-    });
-  }
-};
-
-ChartJS.register(
-  CategoryScale,
-  Colors,
-  LinearScale,
-  PointElement,
-  LineElement,
-  TimeScale,
-  Title,
-  Tooltip,
-  Legend,
-  verticalLinePlugin,
-  annotationsPlugin
-);
+import { DiagnosticsLineChart } from "../shared/diagnostics/line-chart";
 
 type Message = {
   id: string;
@@ -490,7 +350,7 @@ const DiagnosticsResource = ({
                       </div>
                     )}
                     <div className="mt-2 min-h-[200px]">
-                      <SynchronizedHoverLineChartWrapper
+                      <DiagnosticsLineChart
                         showLegend={true}
                         keyId={plot.id}
                         chart={{
@@ -526,204 +386,9 @@ const DiagnosticsResource = ({
   );
 };
 
-const SynchronizedHoverLineChartWrapper = ({
-  showLegend = true,
-  keyId,
-  chart: { labels, datasets: originalDatasets, title },
-  xAxisUnit,
-  yAxisLabel,
-  yAxisUnit,
-  annotations = [],
-  synchronizedHoverContext,
-}: {
-  showLegend?: boolean;
-  keyId: string;
-  chart: {
-    title: string;
-    labels: string[];
-    datasets: Array<{
-      label: string;
-      data: number[];
-    }>;
-  };
-  xAxisUnit: TimeUnit;
-  yAxisLabel?: string;
-  yAxisUnit?: string;
-  annotations?: Annotation[];
-  synchronizedHoverContext: React.Context<HoverState>;
-}) => {
-  const { timestamp, setTimestamp } = useContext(synchronizedHoverContext);
-  const chartRef = React.useRef<ChartJS<"line">>();
-
-  // Truncate sha256 resource names to 8 chars
-  const datasets = originalDatasets.map(dataset => ({
-    ...dataset,
-    label: dataset.label.length === 64 ? dataset.label.slice(0, 8) : dataset.label
-  }));
-
-  if (!datasets || !title) return null;
-
-  React.useEffect(() => {
-    const chart = chartRef.current;
-    if (!chart) return;
-
-    if (!timestamp) {
-      chart.setActiveElements([]);
-      chart.tooltip?.setActiveElements([], { x: 0, y: 0 });
-      chart.update();
-      return;
-    }
-
-    const timestampIndex = labels.indexOf(timestamp);
-    if (timestampIndex === -1) return;
-
-    const activeElements = datasets.reduce<{
-      datasetIndex: number;
-      index: number;
-    }[]>((acc, dataset, datasetIndex) => {
-      if (!dataset.data[timestampIndex]) return acc;
-
-      return [
-        ...acc,
-        {
-          datasetIndex,
-          index: timestampIndex,
-        },
-      ];
-    }, []);
-
-    chart.setActiveElements(activeElements);
-    chart.tooltip?.setActiveElements(activeElements, { x: 0, y: 0 });
-    chart.update();
-  }, [timestamp, labels, datasets]);
-
-  const formatYAxisTick = (value: number, unit?: string) => {
-    if (!unit) return value;
-
-    unit = unit.trim();
-    if (unit === '%') return `${value}%`;
-    if (unit.endsWith('B')) return `${value}${unit}`;
-
-    return value;
-  };
-
-  return (
-    <Line
-      ref={chartRef}
-      datasetIdKey={keyId}
-      data={{
-        labels,
-        datasets,
-      }}
-      options={{
-        responsive: true,
-        maintainAspectRatio: false,
-        animation: false,
-        plugins: {
-          tooltip: {
-            enabled: true,
-            mode: 'index',
-            intersect: false,
-          },
-          colors: {
-            forceOverride: true,
-          },
-          legend: {
-            display: showLegend,
-            labels: {
-              usePointStyle: true,
-              boxHeight: 5,
-              boxWidth: 3,
-              padding: 20,
-            },
-          },
-          title: {
-            font: {
-              size: 16,
-              weight: "normal",
-            },
-            color: "#595E63",
-            align: "start",
-            display: false,
-            text: title,
-            padding: showLegend
-              ? undefined
-              : {
-                top: 10,
-                bottom: 30,
-              },
-          },
-          annotations: annotations,
-        },
-        interaction: {
-          mode: 'index',
-          intersect: false,
-        },
-        onHover: (event, elements, chart) => {
-          if (!event.native) return;
-
-          if (elements && elements.length > 0) {
-            const timestamp = labels[elements[0].index];
-            setTimestamp(timestamp);
-          } else {
-            setTimestamp(null);
-          }
-        },
-        scales: {
-          x: {
-            border: {
-              color: "#111920",
-            },
-            grid: {
-              display: false,
-            },
-            ticks: {
-              color: "#111920",
-              maxRotation: 0,
-              minRotation: 0,
-              autoSkip: true,
-              maxTicksLimit: 5,
-            },
-            adapters: {
-              date: {
-                zone: "UTC",
-              },
-            },
-            time: {
-              tooltipFormat: "yyyy-MM-dd HH:mm:ss 'UTC'",
-              unit: xAxisUnit,
-              displayFormats: {
-                minute: "HH:mm 'UTC'",
-                day: "MMM dd",
-              },
-            },
-            type: "time",
-          },
-          y: {
-            min: 0,
-            border: {
-              display: false,
-            },
-            title: yAxisLabel
-              ? {
-                display: true,
-                text: yAxisLabel,
-              }
-              : undefined,
-            ticks: {
-              callback: (value) => formatYAxisTick(value as number, yAxisUnit),
-              color: "#111920",
-            },
-          },
-        },
-      }}
-    />
-  );
-};
-
 export const DiagnosticsDetailPage = () => {
   // Parse the investigation parameters from the query string.
-  const [searchParams, setSearchParams] = useSearchParams();
+  const [searchParams] = useSearchParams();
   const accessToken = useSelector(selectAccessToken);
   const appId = searchParams.get("appId");
   const symptomDescription = searchParams.get("symptomDescription");
